* runtest.exp: Ensure that multipass pass variables are always 	restored.  Previously, they were only restored if the "Go digging 	for tests" branch was taken near the end of runtest.exp.
This bug was found while splitting a new variable out of some uses of
$srcdir when Emacs paren highlighting did not look right.  Sure
enough, the per-pass variables are set unconditionally, but are only
restored if the "Go digging for tests" branch was taken.  This patch
moves that loop out of an 'else' block and into the same 'foreach'
that contains the loop that sets per-pass variables.

Signed-off-by: Ben Elliston <bje@gnu.org>