Revert "Add hook to put branch and evision in the version for development builds."
This reverts commit 1a76740eff41f63e64f1e65e930ad286f357362c.

This caused problems (reported in PR44462 as a secondary issue)
when a source tree is copied out of a Git directory.  A better
solution will be developed for the 1.6.4 release that also tags
development versions when run from Git working directories without
relying on an installation hook.

The problem can be demonstrated with:
    git -z ls-files | cpio -0dp /tmp/dejagnu-test-src
followed by configuring in that non-Git copy of the source tree.