Remove Tcl version check in configure
Tcl versions old enough to be incompatible with DejaGnu are very unlikely
to be encountered "in the wild" and there were ways for configure to find
one version of Tcl, but DejaGnu to actually run with a different instance
installed somewhere else on the system.

Further, testing for specific versions is generally discouraged in favor
of testing for the features actually used.  Simply running the DejaGnu
testsuite will provide adequate assurance that the Tcl installation
found by the runtest script is sufficient for running DejaGnu.