Remove automatic search for local Expect executable in DejaGnu testsuite
This search supported the old Cygnus tree layout, which has been obsolete
for a very long time.  Anyone still wanting to use a newly-built Expect
to run DejaGnu tests can do so by passing EXPECT explicitly on the runtest
command line (to use the new Expect only for the library unit tests) or in
the environment (which will also cause the runtest launcher script to use
it to run the framework).  Passing EXPECT in the environment can also be
achieved by setting EXPECT on the "make check" command line.

This remediates the test failures in PR47385 but does not address the
underlying root causes, which will be fixed in the 1.6.4 development phase.