Adjust conditional for broken legacy case
DejaGnu has always required testsuites to be rooted in a directory
named exactly "testsuite" but older versions happened to work even
if this requirement was violated.  Special handling for this broken
case was included, but the conditional to invoke this handling was
too narrow and missed several cases in the GNU Automake testsuite.