Revise generation of "END" messages in dejagnu.h
The "END" message is now produced upon normal exit, without requiring that
the totals() function or method be called.  The C++ API now emits totals
only when the last TestState object in the program is destroyed, instead
of every time a TestState object is destroyed.  This required adding code
to track the number of live TestState objects.